Added dij algo, finishing to print the path

+++ Util/graph.h
@@ -7,7 +7,7 @@
 typedef struct arc
 {
   EDiff_t diff; //stores the difficulty of the line
-  unsigned int cost; //cost between this points
+  float cost; //cost between this points
   char * name;
   int indexStart;
   int indexArrival; // the index of the destination
